After file restructure
Original Dev-Somerville branch got out of sync.  So, had to delete it
locally and on server.  Created this new branch and copied over the
project, with the restructured folder directories.  XStyle sub-module is
not included at this point as it might have been causing a lot of the
issues after the restructure- since git still thought the sub-module
should be in the old directory.