Grid resize() bug fix; Styling changes
Any action that resized the Query Results grid, removed the click event
from the zoom Image nodes.  EnhancedGrid lacks eventing for resize() to
reattach the click events.  Implemented on.selector and set the event
handler to the grid itself to fix the issue.

Updated Row header width, and removed Zoom To column name, as well as
shortened the zoom column width.