Data Portal Links in Legend and other fixes
Graphic in legend links to the data portal for metadata and downloading
of the data.  This only works for layers that are in the data portal and
on iMAP's server.

Also including files necessary for export script.  Must set the bin
folder or the whole application as an application in IIS to work.

Some fixes to draw styling.